+++ packages/server_auth/test/options_core_test.dart
@@ -281,4 +281,12 @@ final class _DurableAuthStore implements AuthStore {
   @override
   AuthVerificationTokenStore get verificationTokens =>
       _delegate.verificationTokens;
+
+  @override
+  AuthWebAuthnChallengeStore get webAuthnChallenges =>
+      _delegate.webAuthnChallenges;
+
+  @override
+  AuthWebAuthnAuthenticatorStore get webAuthnAuthenticators =>
+      _delegate.webAuthnAuthenticators;
 }